Added additional constraints and modified user experience.
In the add item feature, a constraint has been added to restrict the user from  saving the item if the information in the input field is incompletely filled up. Search function has been implemented. And the edit and delete features has been modified to show as two pop up options when long pressing a particular item.